feat: 아이템 리스트 조회, 착용 아이템 조회 api 연동
